+++ pkg/app/server/binding/decoder/base_type_decoder.go
@@ -38,21 +38,31 @@
  * Modifications are Copyright 2022 CloudWeGo Authors
  */
 
-package binding
+package decoder
 
 import (
 	"fmt"
 	"reflect"
 
-	"github.com/cloudwego/hertz/internal/bytesconv"
+	path1 "github.com/cloudwego/hertz/pkg/app/server/binding/path"
 	"github.com/cloudwego/hertz/pkg/common/utils"
 )
 
-type mapTypeFieldTextDecoder struct {
-	fieldInfo
+type fieldInfo struct {
+	index       int
+	parentIndex []int
+	fieldName   string
+	tagInfos    []TagInfo    // query,param,header,respHeader ...
+	fieldType   reflect.Type // can not be pointer type
 }
 
-func (d *mapTypeFieldTextDecoder) Decode(req *bindRequest, params PathParam, reqValue reflect.Value) error {
+type baseTypeFieldTextDecoder struct {
+	fieldInfo
+	decoder TextDecoder
+}
+
+func (d *baseTypeFieldTextDecoder) Decode(req *bindRequest, params path1.PathParam, reqValue reflect.Value) error {
+	var err error
 	var text string
 	var defaultValue string
 	for _, tagInfo := range d.tagInfos {
@@ -66,16 +76,25 @@ func (d *mapTypeFieldTextDecoder) Decode(req *bindRequest, params PathParam, req
 		defaultValue = tagInfo.Default
 		if len(ret) != 0 {
 			text = ret[0]
+			err = nil
 			break
 		}
+		if tagInfo.Required {
+			err = fmt.Errorf("'%s' field is a 'required' parameter, but the request does not have this parameter", d.fieldName)
+		}
+	}
+	if err != nil {
+		return err
 	}
 	if len(text) == 0 && len(defaultValue) != 0 {
 		text = defaultValue
 	}
+	//todo: check a=?b=?c= 这种情况 loosemode
 	if text == "" {
 		return nil
 	}
 
+	// get the non-nil value for the field
 	reqValue = GetFieldValue(reqValue, d.parentIndex)
 	field := reqValue.Field(d.index)
 	if field.Kind() == reflect.Ptr {
@@ -88,13 +107,14 @@ func (d *mapTypeFieldTextDecoder) Decode(req *bindRequest, params PathParam, req
 		var vv reflect.Value
 		vv, err := stringToValue(t, text)
 		if err != nil {
-			return fmt.Errorf("unable to decode '%s' as %s: %w", text, d.fieldType.Name(), err)
+			return err
 		}
 		field.Set(ReferenceValue(vv, ptrDepth))
 		return nil
 	}
 
-	err := jsonUnmarshalFunc(bytesconv.S2b(text), field.Addr().Interface())
+	// Non-pointer elems
+	err = d.decoder.UnmarshalString(text, field)
 	if err != nil {
 		return fmt.Errorf("unable to decode '%s' as %s: %w", text, d.fieldType.Name(), err)
 	}
@@ -102,7 +122,7 @@ func (d *mapTypeFieldTextDecoder) Decode(req *bindRequest, params PathParam, req
 	return nil
 }
 
-func getMapTypeTextDecoder(field reflect.StructField, index int, tagInfos []TagInfo, parentIdx []int) ([]decoder, error) {
+func getBaseTypeTextDecoder(field reflect.StructField, index int, tagInfos []TagInfo, parentIdx []int) ([]fieldDecoder, error) {
 	for idx, tagInfo := range tagInfos {
 		switch tagInfo.Key {
 		case pathTag:
@@ -129,7 +149,13 @@ func getMapTypeTextDecoder(field reflect.StructField, index int, tagInfos []TagI
 	for field.Type.Kind() == reflect.Ptr {
 		fieldType = field.Type.Elem()
 	}
-	fieldDecoder := &mapTypeFieldTextDecoder{
+
+	textDecoder, err := SelectTextDecoder(fieldType)
+	if err != nil {
+		return nil, err
+	}
+
+	return []fieldDecoder{&baseTypeFieldTextDecoder{
 		fieldInfo: fieldInfo{
 			index:       index,
 			parentIndex: parentIdx,
@@ -137,7 +163,6 @@ func getMapTypeTextDecoder(field reflect.StructField, index int, tagInfos []TagI
 			tagInfos:    tagInfos,
 			fieldType:   fieldType,
 		},
-	}
-
-	return []decoder{fieldDecoder}, nil
+		decoder: textDecoder,
+	}}, nil
 }
