+++ pkg/app/server/binding/decoder/tag.go
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package binding
+package decoder
 
 import (
 	"reflect"
