+++ pkg/app/server/binding/path/path.go
@@ -1,0 +1,6 @@
+package path
+
+// PathParam parameter acquisition interface on the URL path
+type PathParam interface {
+	Get(name string) (string, bool)
+}
